+++ mvc/src/main/java/nextstep/mvc/DispatcherServlet.java
@@ -1,25 +1,24 @@
 package nextstep.mvc;
 
+import jakarta.servlet.ServletConfig;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import nextstep.mvc.adapter.HandlerAdapter;
+import nextstep.mvc.mapping.HandlerMapping;
+import nextstep.mvc.registry.HandlerAdapterRegistry;
+import nextstep.mvc.registry.HandlerMappingRegistry;
 import nextstep.mvc.view.ModelAndView;
 import nextstep.mvc.view.View;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Objects;
-
 public class DispatcherServlet extends HttpServlet {
 
     private static final long serialVersionUID = 1L;
     private static final Logger log = LoggerFactory.getLogger(DispatcherServlet.class);
-
-
     private final HandlerAdapterRegistry handlerAdapterRegistry;
     private final HandlerMappingRegistry handlerMappingRegistry;
