+++ mvc/src/main/java/nextstep/mvc/adapter/AnnotationHandlerAdapter.java
@@ -2,7 +2,6 @@ package nextstep.mvc.adapter;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.HandlerAdapter;
 import nextstep.mvc.controller.tobe.HandlerExecution;
 import nextstep.mvc.view.ModelAndView;
