+++ mvc/src/main/java/nextstep/mvc/adapter/HandlerAdapter.java
@@ -1,4 +1,4 @@
-package nextstep.mvc;
+package nextstep.mvc.adapter;
 
 import nextstep.mvc.view.ModelAndView;
