+++ mvc/src/test/java/nextstep/mvc/controller/tobe/AnnotationHandlerMappingTest.java
@@ -15,7 +15,7 @@ class AnnotationHandlerMappingTest {
 
     @BeforeEach
     void setUp() {
-        handlerMapping = new AnnotationHandlerMapping("samples");
+        handlerMapping = new AnnotationHandlerMapping( "samples");
         handlerMapping.initialize();
     }
